requirements.txt
certifi==2020.12.5
chardet==4.0.0
docker==4.4.1
idna==2.10
paho-mqtt==1.5.1
requests==2.25.1
six==1.15.0
urllib3==1.26.2
websocket-client==0.57.0
